Let the Errors Flow.
If a user enters bad data, we shouldn' just blow up with a 500. We should kindly tell them, and then swiftly reject
their request with a 400 - Dikembe Mutombo style.

Also, if Tax Jar isn't happy with the request, its probably because the user passed in bad data. In this case, we should
allow the response status code and error `details` to flow from Tax Jar to the caller so they can take whatever action they
choose. After all, we are just the middle man. Don't shoot the messenger.

Note, there is a sort of code smell in the error handling for condition 1 above to make it match the *shape* of condition 2.
This makes the try catch easier to reason about and deal with... Having said that, this is a great example of why `Either`s are
so wonderful. They take this code smell away -- yes, of course there is a trade off of being more explicit about your errors.
May not be worth it in this trivial example, but a good example of how `Either`s can help keep things tidy.

Finally, include the `toAddress` and `fromAddress` in the api response. This is offered as a convenience to the caller (and
for demo purposes).